Trait inexor_rgf_reactive_model_api::reactive_property_container::ReactivePropertyContainer

pub trait ReactivePropertyContainer {
    // Required methods
    fn tick_checked(&self);
    fn tick(&self);
    fn has_property(&self, name: &str) -> bool;
    fn add_property<S: Into<String>>(
        &self,
        name: S,
        mutability: Mutability,
        value: Value
    );
    fn add_property_by_type(&self, property: &PropertyType);
    fn remove_property<S: Into<String>>(&self, name: S);
    fn observe_with_handle<F>(&self, name: &str, subscriber: F, handle_id: u128)
       where F: FnMut(&Value) + 'static + Send;
    fn remove_observer(&self, name: &str, handle_id: u128);
    fn remove_observers(&self, name: &str);
    fn remove_all_observers(&self);
}

Required Methods

fn tick_checked(&self)

Sends the current value down the stream if mutable.

fn tick(&self)

Sends the current value down the stream.

fn has_property(&self, name: &str) -> bool

Returns true, if a property with the given name exists.

fn add_property<S: Into<String>>( &self, name: S, mutability: Mutability, value: Value )

Adds a reactive property with the given name and the given initial value.

fn add_property_by_type(&self, property: &PropertyType)

Adds a reactive property with the given name and the given initial value.

fn remove_property<S: Into<String>>(&self, name: S)

Removes the reactive property with the given name.

fn observe_with_handle<F>(&self, name: &str, subscriber: F, handle_id: u128)
where F: FnMut(&Value) + 'static + Send,

Observe the stream output flowing out of the stream of the property with the given name. The handle_id allows to remove the observer again.

fn remove_observer(&self, name: &str, handle_id: u128)

Removes the subscriber with the given handle_id from the stream of the property with the given name.

fn remove_observers(&self, name: &str)

Removes the subscribers of the property with the given name.

fn remove_all_observers(&self)

Removes all subscribers of all properties.

Object Safety

This trait is not object safe.